Keeping house clean, Only time when i doing it is like this

+++ Fraples7DevDX3D/Source/Fraples7DevDX3D/Core/Application.cpp
@@ -119,7 +119,7 @@ namespace FraplesDev
 		{
 			if (!_mWin.IsCursorEnabled())
 			{
-				_mCamera.Rotate(delta->x, delta->y);
+				_mCamera.Rotate((float)delta->x, (float)delta->y);
 			}
 
 		}
